pyproject.toml
[project]
name = "clawmark"
version = "0.1.0"
description = "ClawMark — Multimodal AI Agent Benchmark"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "playwright>=1.40",
    "rich>=13.0",
    "google-api-python-client>=2.100",
    "google-auth>=2.25",
    "caldav>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
clawmark = "clawmark.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clawmark"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pillow>=12.1.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
